pyproject.toml
[project]
name = "Lareta"
version = ""
description = ""
authors = [
    {name = "BlueGlassBlock", email = "blueglassblock@outlook.com"},
    {name = "GreyElaina", email = "GreyElaina@outlook.com"},
    {name = "ProgramRipper", email = "programripper@foxmail.com"},
    {name = "RF-Tar-Railt", email = "3165388245@qq.com"},
]
dependencies = [
    "graia-ariadne[full] @ https://hub.fastgit.xyz/GraiaProject/Ariadne/archive/refs/heads/dev.zip",
    "graia-broadcast @ https://hub.fastgit.xyz/GraiaProject/BroadcastControl/archive/refs/heads/master.zip",
    "sqlmodel>=0.0.6",
    "aiosqlite>=0.17.0",
    "creart-graia>=0.1.4",
]
requires-python = ">=3.10"
license = {text = "AGPL-3.0-only"}
[project.optional-dependencies]

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.3.0",
    "isort>=5.10.1",
]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
